requirements.txt
flask==2.0.1
tweepy==4.10.0
praw==7.6.0
nltk==3.7
plotly==5.10.0
pandas==1.4.2
Flask-SocketIO==5.1.1
python-dotenv==0.20.0
werkzeug==2.0.1
# Deployment requirements
gunicorn==20.1.0
eventlet==0.33.3 